+++ tests/Converters/ShopsConverterTest.php
@@ -30,27 +30,32 @@ class ShopsConverterTest extends PHPUnit_Framework_TestCase
                 {
                   "id": 4092,
                   "name": "da capo - katarynki i Gracze",
-                  "products_url": "/sklep:katarynki-jasky-pl.html"
+                  "products_url": "/sklep:katarynki-jasky-pl.html",
+                  "url_logo": "/s/632-201412241238343.jpg"
                 },
                 {
                   "id": 8451,
                   "name": "dabar.pl",
-                  "products_url": "/sklep:dabar-pl.html"
+                  "products_url": "/sklep:dabar-pl.html",
+                  "url_logo": "/s/632-201412241238342.jpg"
                 },
                 {
                   "id": 21827,
                   "name": "Dadin Maluchy to kochają",
-                  "products_url": "/sklep:dadin-dadin-pl.html"
+                  "products_url": "/sklep:dadin-dadin-pl.html",
+                  "url_logo": "/s/632-201412241238341.jpg"
                 },
                 {
                   "id": 5332,
                   "name": "Dareckishop.eu",
-                  "products_url": "/sklep:dareckishop-eu.html"
+                  "products_url": "/sklep:dareckishop-eu.html",
+                  "url_logo": "/s/632-201412241238345.jpg"
                 },
                 {
                   "id": 9338,
                   "name": "DaroTrans",
-                  "products_url": "/sklep:darotrans-pl.html"
+                  "products_url": "/sklep:darotrans-pl.html",
+                  "url_logo": "/s/632-20141224123834.jpg"
                 }
               ]
             }