+++ CHANGELOG.md
@@ -7,6 +7,7 @@ master
 - zmiana interface \Nokaut\ApiKit\Cache\CacheInterface oraz dodanie Nokaut\ApiKit\Cache\AbstractCache - jeśli implementowaliście interface CacheInterface wystarczy, że rozszerzycie swoją klasę abstrkcyjną klasą Nokaut\ApiKit\Cache\AbstractCache
 - Pobieranie ShopsRepository i ProducersRepository z klasy ApiKit
 - Dodanie repozytoriów: ShopsAsyncRepository i ProducersAsyncRepository
+- Entity/Shop dodanie url_logo
 
 v1.6.0
 ------